ir: dedup test fn rb into bound_box_from_opref helper
Add a `test-support` feature to majit-ir that exposes
`box_ref::bound_box_from_opref`, which binds an OpRef to a bound BoxRef
(fresh InputArg / SameAs* Op at the same position, parked in a
thread-local producer pool; None/Const route through from_opref).

Replace the nine byte-identical `fn rb` copies in the backend, gc, and
jit-trace test modules with `use bound_box_from_opref as rb`, and wire
`majit-ir = { features = ["test-support"] }` into the dev-dependencies
of the five consumer crates.
