+++ majit/majit-backend-dynasm/Cargo.toml
@@ -16,4 +16,5 @@ dynasmrt = { workspace = true }
 libc = { workspace = true }
 
 [dev-dependencies]
+majit-ir = { workspace = true, features = ["test-support"] }
 majit-trace = { workspace = true }
