+++ majit/majit-backend-wasm/Cargo.toml
@@ -30,5 +30,6 @@ wasm-encoder = { workspace = true }
 wasm-bindgen = { workspace = true, optional = true }
 
 [dev-dependencies]
+majit-ir = { workspace = true, features = ["test-support"] }
 wasmparser = { workspace = true }
 smallvec = { workspace = true }
