+++ majit/majit-gc/Cargo.toml
@@ -26,3 +26,6 @@ majit-ir = { workspace = true }
 [target.'cfg(target_os = "macos")'.dependencies]
 # env.py:413-433 get_L2cache_darwin uses sysctl() to size the nursery.
 libc = { workspace = true }
+
+[dev-dependencies]
+majit-ir = { workspace = true, features = ["test-support"] }
